+++ src/TAO/Operation/include/operations.h
@@ -166,7 +166,7 @@ namespace TAO
          *
          **/
         bool Trust(const uint256_t &hashAddress, const uint1024_t &hashLastTrust,
-            const uint32_t nSequence, const uint32_t nLastTrust, const uint64_t nAmount, const uint256_t &hashCaller,
+            const uint32_t nSequence, const uint64_t nLastTrust, const uint64_t nAmount, const uint256_t &hashCaller,
             const uint8_t nFlags, TAO::Ledger::Transaction &tx);
